+++ src/js/presentaciones/cobrar-presentacion/components/toolbar/ReciboInput.js
